codegen: Fix various pointer/integer size mismatches
In several places in codegen we assume that pointers are always represented as
64-bit integers. This is no longer the case (to support 32-bit architectures),
so type mismatches can occur when generating the IR.

For example, when casting an int to a pointer ("$a = (uint32 *) 0xdeadbeef") on
armv7, $a is 4-bytes (pointer type) but the RHS evaluates to an int64, so we end
up with the following invalid IR:

  # bpftrace --emit-llvm test -e 'BEGIN { $a = (uint32 *) 0xdeadbeef }'
  $ llc test.original.ll
  llc: error: llc: test.original.ll:15:9: error: stored value and pointer type do not match
    store i64 3735928559, i32* %"$a", align 4
          ^

This patch adds integer casts to places where such mismatches happen and replaces
a few getInt64(Ty) calls with pointer-sized variants. Note that this should be a
no-op on 64-bit architectures.